requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
hidapi==0.9.0.post2
idna==2.9
pkg-resources==0.0.0
python-engineio==3.12.1
python-socketio==4.5.1
requests==2.23.0
six==1.14.0
urllib3==1.25.9